% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{bad_code_copies}
\alias{bad_code_copies}
\title{Copies of rows with invalid codes}
\usage{
bad_code_copies(dat, checked_list)
}
\arguments{
\item{dat}{Original data object used on the function \code{check_codes}}

\item{checked_list}{The list returned after using \code{check_codes}}
}
\value{
a list with subsets of data corresponding to rows with invalid codes
}
\description{
Copies of rows with invalid codes
}
\examples{
# see check_codes first, after that do:
bad_copies <- bad_code_copies(dat, codes_checked)
}

