% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{bad_ts_copies}
\alias{bad_ts_copies}
\title{Copies of rows with invalid timestamps}
\usage{
bad_ts_copies(dat, checked_list)
}
\arguments{
\item{dat}{Original data object used on the function \code{check_timestamps}}

\item{checked_list}{The list returned after using \code{check_timestamps}}
}
\value{
a list with subsets of data corresponding to rows with invalid timestamps
}
\description{
Copies of rows with invalid timestamps
}
\examples{
# see check_timestamps first, after that do:
bad_timestamps <- bad_ts_copies(dat, ts_checked)
}

