% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{check_codes}
\alias{check_codes}
\title{Check columns for bad codes}
\usage{
check_codes(code_list, dat, as.na = TRUE)
}
\arguments{
\item{code_list}{A list of column names with each name having a vector of codes to check.}

\item{dat}{The data.frame/data.table to check}

\item{as.na}{If \code{TRUE}, will return a new data set with \code{NA} instead of the bad code.}
}
\value{
A list containing two items. \code{$data} is the new data with NAs, 
\code{$bad_codes} is another list, each item corresponding to the input list. In 
each there is a data.frame that is either empty (no bad codes found), or contains indices
for the row numbers with bad codes and the column name and type of bad code found. 
If \code{as.na=FALSE}, the new data will be \code{NULL}, and will only return
\code{$bad_codes}, if any.
}
\description{
Check for invalid codes used for each column and argument specified
}
\details{
This takes an already existing data.frame/data.table and checks specific columns 
to see if they contain only the codes listed in \code{code_list}. Each item in the 
list must be named according to the column name in the data.frame, and each item
must contain a vector of 1 or more of valid codes to check. Codes can be numeric or
characters.
}
\examples{
# Use example data
dat <- ordinal_align()

# Make the list of valid codes, names corresponding to columns in the data
code_list <- list(
    childhands.hand = c("left", "right", "both"),
    childhands.look = c(0,1),
    parenthands.hand = c("left", "right", "both"),
    parenthands.look = c(0,1)
)

# check for bad codes, returning new data with bad codes as NAs
codes_checked <- check_codes(code_list, dat)
}

