% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{check_timestamps}
\alias{check_timestamps}
\title{Check for invalid timestamps}
\usage{
check_timestamps(ts_list, dat, tmin = 0, tmax = 86400000, as.na = TRUE)
}
\arguments{
\item{ts_list}{A list with each item corresponding to either a pair of timestamp columns
\code{c(onset, offset)}, or just a single column name. If a single name, durations will not be checked,
and only out of range will be checked.}

\item{dat}{A data.frame/data.table that contains the columns with the onset/offset timestamps}

\item{tmin}{Minimum allowed timestamp. Defaults to 0 (milliseconds)}

\item{tmax}{Maximum allowed timestamp. Defaults to one day in milliseconds.}

\item{as.na}{Set bad timestamps to \code{NA}}
}
\value{
Returns a list with 3 items, \code{$data, $ranges, $durations}. 
\code{$data} is new data with bad timestamps as \code{NA}. \code{$ranges} is
a list of column names, with a vector of indices corresponding to timestamps
out of range. If all okay, the vector is empty. \code{$durations} is a list
of bad timestamp durations. If \code{as.na=TRUE}, the offset timestamp is 
set to \code{NA}.
}
\description{
Check for invalid timestamps
}
\examples{
# Use example data
dat <- ordinal_align()

# Make the list of timestamps to check in the example data
# These check durations as well as out of range
ts_list <- list(
    c(on="parenthands.onset", off="parenthands.offset"), # explicit
    child = c("childhands.onset", "childhands.offset") # infer from order which is on/off
)
ts_checked <- check_timestamps(ts_list, dat)

# This only checks for bad ranges since its only one column name.
# ts_list can be mixed with 2 or 1 items each.
ts_list <- list("parenthands.offset")
ts_checked2 <- check_timestamps(ts_list, dat)
}

