% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv2datavyu.R
\name{compile_dvcolumn}
\alias{compile_dvcolumn}
\title{Import .csv to R}
\usage{
compile_dvcolumn(folder, column, ..., asList = FALSE)
}
\arguments{
\item{folder}{Character string of the name of the folder to be scanned.}

\item{column}{The name of the column to import as used in the Datavyu file.}

\item{...}{Additional options passed to the read.csv function}
}
\description{
Import Datavyu column to R when using the datavyu2csv.rb script
}
\details{
If the same column name was used but has different number of arguments then you will get an error unless asList=FALSE.
This function assumes that the .csv is structured in a way based on how the datavyu2csv.rb script exports data.
}
\examples{
scan_colnames("myfolder")
}

