% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{datavyu-alignment}
\alias{datavyu-alignment}
\alias{ordinal_align}
\alias{temporal_align}
\alias{horz_merge_datavyu_list}
\title{Merge from a list horizontally}
\usage{
ordinal_align(...)

temporal_align(fps = 30, keep_frame_num = TRUE, ...)

horz_merge_datavyu_list(.list, .f, ..., fps = NULL)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[=import_datavyu_to_list]{import_datavyu_to_list()}}}

\item{fps}{a common framerate to use for alignment. Only set if you want to align
temporally. If not set then it will default to ordinal alignment. For
\code{temporal_align} this defaults to 30 FPS.}

\item{keep_frame_num}{Keep the frame_number column, which will result in a \strong{much} larger
data set. If \code{keep_frame_num=FALSE}, then all you know is which two events overlap
to some degree, not knowing for how long. If \code{keep_frame_num=TRUE} (default), you
can calculate the number of frames that overlap between two events, and if fps is
know, you can convert this total to milliseconds.
video.}

\item{.list}{a nested list returned from \code{\link[=import_datavyu_to_list]{import_datavyu_to_list()}}}

\item{.f}{a function to apply to each data set before anything is done to the data.}
}
\value{
a data.table with aligned data.
}
\description{
Merge a list of imported data temporally or ordinally. This aligns data horizontally
by appending and merging new columns to a common data set.
}
\section{Functions}{
\itemize{
\item \code{ordinal_align}: Align and merge data by cell number (ordinal
alignment). Having a common ordinal number across Datavyu columns may not have much
meaning, but this will allow you to append columns within a single data set.

\item \code{temporal_align}: This will align several Datavyu columns using a
common framerate. You have the option to set the framerate using \code{fps}. The lower
the frame rate, the less likely two events will line up in time. This is because
the timestamps are converted to frame numbers based on chunking them into discrete bins. The
larger the fps, the larger the bins, and more likely two events will line up in
time.

\item \code{horz_merge_datavyu_list}: Merge a list of imported data temporally or ordinally
and align horizontally by appending new columns to a common data set.
}}

\examples{
# set folder path if needed, otherwise use default path with example data.
# options(datavyur.folder="mydatafolder")

#-- Align by a common ordinal number
ord_aligned <- ordinal_align()


# -- Align by a common frame number

# example data with no arguments
ex_data_aligned <- temporal_align(fps=10)

ex_data_aligned_no_frames <- temporal_align(fps=10, keep_frame_num=FALSE)

# example data selecting only one of the columns
ex_data_aligned2 <- temporal_align(columns="childhands")


# -- Align by ordinal or frame but from a list of data objects.
.list <- import_datavyu_to_list()
ord_aligned <- horz_merge_datavyu_list(.list)
}
