% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dvfun.R
\name{datavyu_col_search}
\alias{datavyu_col_search}
\title{Scan .csv files for data}
\usage{
datavyu_col_search(folder, unq = FALSE, cname = "column")
}
\arguments{
\item{folder}{Character string of the name of the folder to be scanned.}

\item{unq}{Return only unique column names}

\item{cname}{Name of .csv column to check if exists. Defaults to \code{"column"}.}
}
\description{
Scans data exported as .csv from Datavyu and returns a list of column names
}
\details{
The function will read all csv files found in the folder and check if they have the column name as indicated in the \code{cname} argument.
If the column name is found, it will return the name of the actual column name used within Datavyu.
}
\examples{
datavyu_col_search("myfolder")
}

