% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavyur.R
\docType{package}
\name{datavyur}
\alias{datavyur}
\title{datavyur: An R Package for managing exported Datavyu files}
\description{
A set of R functions to easily manage coded data exported
from the Datavyu software
}
\details{
To see the documentation on how to use this package, run the following
code in the console:
\cr\cr \code{vignette("tutorial", package="datavyur")} \cr\cr
This will pull up the R vignette "tutorial" and demonstrate
how to use the set of exported functions.
}
