% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{datavyur_data_folder}
\alias{datavyur_data_folder}
\title{Return the data folder for the \code{datavyur} package}
\usage{
datavyur_data_folder()
}
\value{
path string
}
\description{
This is the default folder if none is specified in the R option "datavyur.folder"
}
\examples{
datavyur_data_folder()
}
