% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavyur_example.R
\name{datavyur_example}
\alias{datavyur_example}
\title{Load Sample Datavyu File}
\usage{
datavyur_example(path = "9807_117035.opf")
}
\arguments{
\item{path}{A character string. The file path to the sample Datavyu file
distributed with the datavyur package. Default is ""9807_117035.opf", a
Datavyu file from volume 1, session 9807, with asset_id 117035.}
}
\value{
The full file path to the file.
}
\description{
Load Sample Datavyu File
}
\examples{
datavyur_example() # Shows path to sample Datavyu file.

}
