% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_opf.R
\name{extract_opf}
\alias{extract_opf}
\title{Extracts Elements From Datavyu coding file.}
\usage{
extract_opf(
  in_fn = NULL,
  out_dir = tempdir(),
  create_dir = FALSE,
  auto_write_over = TRUE,
  vb = FALSE
)
}
\arguments{
\item{in_fn}{A character string string. The full file name for the Datavyu
file. Default is NULL.}

\item{out_dir}{A character string. Output directory to save exported files.}

\item{create_dir}{A logical value. Create a new directory. Default is FALSE.}

\item{auto_write_over}{A logical value. If TRUE, any existing files in
out_dir are overwritten. Default is TRUE.}

\item{vb}{A logical value. Provide verbose output. Default is FALSE.}
}
\value{
The output directory where the extracted files were saved.
}
\description{
Extracts Elements From Datavyu coding file.
}
\examples{
\donttest{
extract_opf()
}

}
