% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{fake_datavyu_data}
\alias{fake_datavyu_data}
\title{Fake Datavyu data}
\usage{
fake_datavyu_data(n1 = 10L, n2 = 15L)
}
\arguments{
\item{n1}{Sample size for variable 1}

\item{n2}{Sample size for variable 2}
}
\value{
List of datavyu formated data
}
\description{
This function will create fake data in the R format needed to import back into the
Datavyu software
}
\details{
The function shows how you can have either a list of columns or an actual data frame.
Either way will work.
}
\examples{
my_data <- fake_datavyu_data()
}
