% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{framerate_expand}
\alias{framerate_expand}
\title{Expand a data set containing onset/offset columns to reflect their frame numbers.}
\usage{
framerate_expand(.data, fps, keep_ts = TRUE)
}
\arguments{
\item{.data}{A \code{data.frame} with \code{onset} and \code{offset} columns.}

\item{fps}{The frames per second corresponding to the coded video.}

\item{keep_ts}{Whether to keep the \code{onset} and \code{offset} columns in the data.}
}
\value{
A \code{data.table} with the added column \code{frame_number}
}
\description{
Expand a data set containing onset/offset columns to reflect their frame numbers.
}
\examples{
my_data <- data.frame(onset=1000, offset=1999, code="hello")
framerate_expand(my_data, 30)

my_data2 <- data.frame(onset=c(-1, 0), offset=c(100, 125), x=c(".", "-"))
framerate_expand(my_data2, 30)
}
