% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dvfun.R
\name{import_column}
\alias{import_column}
\title{Import Datavyu column to R}
\usage{
import_column(column, folder = getOption("datavyur.folder"), asList = FALSE,
  ...)
}
\arguments{
\item{column}{The name of the column to import as used in the Datavyu file}

\item{folder}{Character string of the name of the folder to be scanned. Defaults to option \code{datavur.folder}.}

\item{asList}{Logical value indicating to return a list or data frame}

\item{...}{Additional options passed to the read.csv function}
}
\description{
Imports a Datavyu column to R when using the datavyu2csv.rb script
}
\details{
This function only works if you had previously used the \code{datavyu2csv.rb} script to export a Datavyu file to .csv
This can be obtained from \url{http://github.com/iamamutt/datavyu/general}.

Note: If the same column name was used but has different number of arguments then you will get an error unless \code{asList=TRUE}.
This function assumes that the .csv is structured in a way based on how the \code{datavyu2csv.rb} script exports data.
}
\examples{
import_column("myfolder", "mycolumn")
}

