% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{import_file}
\alias{import_file}
\title{Import datavyu data from file name(s)}
\usage{
import_file(file, folder = getOption("datavyur.folder"), asList = FALSE,
  append.colnames = TRUE, classes = getOption("datavyur.classlist"), ...)
}
\arguments{
\item{file}{A character vector of file names}

\item{folder}{Character string corresponding to the folder path to be scanned. Defaults to option \code{datavyur.folder}.}

\item{asList}{Logical value indicating to return a list or data frame (default).}

\item{append.colnames}{If \code{true}, add column name to each argument, 
e.g., \code{column.arg}, instead of having column as a variable in the data.}

\item{classes}{List of new classes to override guessed classes when reading in .csv data. see \code{data.table::fread}}

\item{...}{Additional options passed to the function \code{data.table::fread}}
}
\value{
A list or data.frame
}
\description{
Import datavyu data from file name(s)
}

