% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{is_NA}
\alias{is_NA}
\title{Check for \code{NA}'s but not \code{NaN}'s}
\usage{
is_NA(x)
}
\arguments{
\item{x}{atomic numeric vector}
}
\value{
logical vector
}
\description{
\code{NaN} values are used to denote missing values from the original Datavyu data.
}
\examples{
is_NA(c(0, NaN, NA, 1.0))
}
