% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_as_df.R
\name{load_as_df}
\alias{load_as_df}
\title{Load A Raw Datavyu File as a Data Frame}
\usage{
load_as_df(dv_fn = datavyur_example(), vb = FALSE)
}
\arguments{
\item{vb}{A logical value. Show verbose output. Default is FALSE.}

\item{df_fn}{A character string. The full file name for the Datavyu file to
load. Default is the value returned by \code{datavyur_example()}.}
}
\description{
Load A Raw Datavyu File as a Data Frame
}
