% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dvfun.R
\name{merge_by_time}
\alias{merge_by_time}
\title{Merge by timestamp}
\usage{
merge_by_time(x, y, ids = c(".1", ".2"), keepall = TRUE, mergeby = NULL)
}
\arguments{
\item{x}{top level data frame, (e.g., trial or block)}

\item{y}{lower level data frame (e.g., eye gazes within trial)}

\item{ids}{Suffixes to use to identify repeated column names}

\item{keepall}{whether to keep all non-matching rows (true) or throw away non-matching (false).}

\item{mergeby}{any additional common columns to merge by. Defaults to \code{file}.}
}
\description{
Merges two data frames by onset/offset timestamps
}
\details{
If data is nested, then this will repeat rows from the higher level data. Both x and y must be data frames.
}
\examples{
# get data with some rows not nested
x <- as.data.frame(datavyu_dat(n1=25, n2=2)[[1]])
y <- datavyu_dat(n1=2, n2=100)[[2]]

z1 <- merge_by_time(x, y)
z2 <- merge_by_time(x, y, ids=c(".higher", ".lower"), keepall=FALSE)
}

