% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{merge_nested}
\alias{merge_nested}
\title{Merge nested data}
\usage{
merge_nested(outer_col, inner_col, ...)
}
\arguments{
\item{outer_col}{top level column name as a character, (e.g., trial or block)}

\item{inner_col}{bottom level column name as a character (e.g., eye gazes within
trial)}

\item{...}{additional parameters passed to \code{\link[=import_datavyu]{import_datavyu()}}.}
}
\description{
Merges two Datavyu columns by onset/offset timestamps
}
\details{
Since data is nested, this will repeat rows from the higher level data. If any cell
in the lower level is not within a cell in the higher level, it will not be in the
final data set. This function is to be used only with truly nested data. For finer
control over potentially overlapping cells see \code{\link[=temporal_align]{temporal_align()}} with the
option \code{keep_frame_num=FALSE} and a high value for \code{fps} for better precision
in aligning timestamps.
}
\examples{
\dontrun{
# merge nested data, throwing away lower level cells that
# are not within higher level cells
merged <- merge_nested("childhands", "parenthands")

# another way to merge nested, but contains NAs for cells that are not
# fully within one another since these data are not nested, you can see how much
# is being thrown away compared to x
y <- temporal_align(
  columns=c("childhands", "parenthands"),
  fps=30,
  keep.frames=FALSE
)
}
}
