% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{merge_opf_files_for_export}
\alias{merge_opf_files_for_export}
\title{Combine multiple files into a single list}
\usage{
merge_opf_files_for_export(file, merge.cells = TRUE,
  folder = getOption("datavyur.folder"),
  classes = getOption("datavyur.classlist"), ...)
}
\arguments{
\item{file}{A character vector of file names}

\item{merge.cells}{If the same column names exist, you can merge cells or set as two different cols.}

\item{folder}{Character string corresponding to the folder path to be scanned. Defaults to option \code{datavyur.folder}.}

\item{classes}{List of new classes to override guessed classes when reading in .csv data. see \code{data.table::fread}}

\item{...}{Additional options passed to the function \code{data.table::fread}}
}
\value{
A list for export
}
\description{
Combine multiple files into a single list
}
\examples{
files_to_merge <- c("dyad1", "dyad2", "dyad3")
export_list <- merge_opf_files_for_export(files_to_merge)
r2datavyu(export_list, "merged_file_name")
}

