% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{ms2time}
\alias{ms2time}
\title{Convert milliseconds to a time string}
\usage{
ms2time(timestamps)
}
\arguments{
\item{timestamp}{A numeric time duration in \strong{milliseconds}, such as \code{6000}ms
representing six seconds.}
}
\description{
This will take a duration of time in milliseconds and convert it to the time string
format used by Datavyu.
}
\examples{
# 1102013 = 18 minutes and 22 seconds and 13 milliseconds
ms2time(1102013)

timestamps <- c(0, 300000, 86399999)
ms2time(timestamps)
}
