% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{ruby_script_folder}
\alias{ruby_script_folder}
\title{Return the path to the ruby scripts used within the Datavyu software}
\usage{
ruby_script_folder()
}
\value{
path string
}
\description{
Return the path to the ruby scripts used within the Datavyu software
}
\examples{
ruby_script_folder()
}
