% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv2datavyu.R
\name{scan_colnames}
\alias{scan_colnames}
\title{Scan .csv files for data}
\usage{
scan_colnames(folder, unq = FALSE)
}
\arguments{
\item{folder}{Character string of the name of the folder to be scanned.}

\item{unq}{Return only unique column names found.}
}
\description{
Scans data exported as .csv from Datavyu and returns a list of file paths
}
\examples{
scan_colnames("myfolder")
}

