% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dvfun.R
\name{temporal_align}
\alias{temporal_align}
\title{Temporally align data}
\usage{
temporal_align(dlist, folder = NULL)
}
\arguments{
\item{dlist}{Character vector  to import or list of data frames to merge}

\item{folder}{Defaults to option}
}
\description{
Soon
}
\details{
Soon
}
\examples{
soon
}

