% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{ts2frame}
\alias{ts2frame}
\title{Convert timestamps to frame numbers}
\usage{
ts2frame(x, fps = 30, tstart = 0L, tend = NULL, as_ts = FALSE, warn = TRUE)
}
\arguments{
\item{x}{Vector of timestamps in \strong{milliseconds}}

\item{fps}{Frames per second of the video source. Defaults to 30 fps. The smaller the
value, the more likely two events will be binned together and marked as the same
frame number.}

\item{tstart}{Start timestamp in milliseconds (optional). Any timestamp before
\code{start} start will be \code{NA}, and the frame number will start at \code{1} where above
\code{tstart}. A warning will be triggered if anything outside of \code{tstart} and \code{tend} is
found, unless \code{warn=FALSE}.}

\item{tend}{Longest possible timestamp in milliseconds (optional). Any timestamp past
\code{tend} will be \code{NA}. Defaults to max of \code{x} if not used.}

\item{as_ts}{If set to \code{TRUE}, will return a timestamp representation of the frame
number, adjusted for \code{tstart} and \code{tend}.}

\item{warn}{Turn on/off warnings if any \code{NA}s were found.}
}
\value{
A numeric vector of the same size as \code{x}.
}
\description{
This is typically used for aligning cells by a common time.
}
\details{
Frame numbers may be repeated if several timestamps are close in duration within the
specified frames per second.
}
\examples{
t <- c(
  -1000, 100, 4047, 7451, 14347, -1,
  27920, 30669, 39798, 42504, 42510, 51451, 56034
)
ts2frame(t, tstart=1000, warn=FALSE)
}
