% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{vert_merge_datavyu_list}
\alias{vert_merge_datavyu_list}
\title{Merge from a list vertically}
\usage{
vert_merge_datavyu_list(.list, .f, ...)
}
\arguments{
\item{.list}{a nested list returned from \code{\link[=import_datavyu_to_list]{import_datavyu_to_list()}}}

\item{.f}{a function to apply to each data set before anything is done to the data.}

\item{...}{additional arguments passed to \code{.f}}
}
\value{
a data.table with stacked data.
}
\description{
This will melt all data to have common column names and convert all data to
characters except for ordinal, onset, offset.
}
\examples{
.list <- import_datavyu_to_list()
stacked_data <- vert_merge_datavyu_list(.list)
}
