% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{datavyu_col_search}
\alias{datavyu_col_search}
\title{Scan .csv files for data}
\usage{
datavyu_col_search(folder = getOption("datavyur.folder"), unq = FALSE,
  cnames = c("file", "column", "codes", "classes", "local"))
}
\arguments{
\item{folder}{Character string of the name of the folder to be scanned. Defaults to option \code{datavyur.folder}.}

\item{unq}{Return only unique column names}

\item{cnames}{Name of columns to return if \code{unq=TRUE}.}
}
\value{
data.frame with .csv info. The variable \code{file} is the original .opf file name used.
\code{column} is the name of the column within the .opf file. \code{codes} are the argument names used.
\code{classes} are the guessed classes for each argument based on R import functions.
\code{local} is the .csv file name and location for a particular column and .opf file.
}
\description{
Scans data exported as .csv from Datavyu and returns a list of .opf column names among other attributes
}
\details{
The function will read all .csv files found in \code{folder}.
All valid .opfs will be found based on if they have the columns \code{file, column, onset, offset, ordinal}.
}
\examples{
datavyu_col_search("myfolder")
}

