% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{datavyu_dat}
\alias{datavyu_dat}
\title{Fake Datavyu data}
\usage{
datavyu_dat(n1 = 10, n2 = 15)
}
\arguments{
\item{n1}{Sample size for variable 1}

\item{n2}{Sample size for variable 2}
}
\value{
List of datavyu formated data
}
\description{
This function will create fake data in the R format needed to import back into the Datavyu software
}
\details{
The function shows how you can have either a list of columns or an actual data frame.
Either way will work.
}
\examples{
my_data <- datavyu_dat()
}

