% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavyur.R
\docType{package}
\name{datavyur}
\alias{datavyur}
\alias{datavyur-package}
\title{datavyur: An R Package for managing exported Datavyu files}
\description{
A set of R functions to easily manage coded data exported 
from the Datavyu software
}
\details{
To see the documentation on how to use this package, run the following
code in the console:
\cr\cr \code{vignette("tutorial", package="datavyur")}. \cr\cr
This will pull up the R vignette "tutorial" and demonstrate 
how to use the set of exported functions below.
}
\section{Exported functions}{

\code{\link{check_codes}} \cr
\code{\link{check_timestamps}} \cr
\code{\link{datavyu_col_search}} \cr
\code{\link{datavyu_dat}} \cr
\code{\link{import_column}} \cr
\code{\link{merge_nested}} \cr
\code{\link{ms2time}} \cr
\code{\link{multi_merge}} \cr
\code{\link{ordinal_align}} \cr
\code{\link{r2datavyu}} \cr
\code{\link{temporal_align}} \cr
\code{\link{ts2frame}} \cr
\cr \cr
For more details, see \url{https://github.com/iamamutt/datavyu}
}

