% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datavyur.R
\docType{package}
\name{datavyur}
\alias{datavyur}
\alias{datavyur-package}
\title{datavyur: An R Package for dealing with Datavyu files}
\description{
A set of functions to handle import/export between R and Datavyu
}
\section{Exported functions}{

\code{\link{datavyu_col_search}} \cr
\code{\link{datavyu_dat}} \cr
\code{\link{import_column}} \cr
\code{\link{merge_nested}} \cr
\code{\link{ms2time}} \cr
\code{\link{multi_merge}} \cr
\code{\link{ordinal_align}} \cr
\code{\link{r2datavyu}} \cr
\code{\link{temporal_align}} \cr
\code{\link{ts2frame}} \cr
}

