% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codes.R
\name{extract_codes}
\alias{extract_codes}
\title{Extract Codes From a Datavyu file.}
\usage{
extract_codes(
  in_dir = ".",
  dv_fn = "db",
  out_dir = in_dir,
  out_fn = paste0(out_dir, "/", "codes.csv"),
  auto_write_over = TRUE,
  code_regex = "^([a-zA-Z_]+[0-9]*[a-zA-Z_]*[0-9]*)",
  code_vals_regex = "\\\\)-([a-zA-Z\\\\/]+)\\\\|",
  code_type_regex = "([a-zA-Z]+)$",
  vb = FALSE
)
}
\arguments{
\item{in_dir}{A character string. Input directory. Defaults to '.'}

\item{dv_fn}{A character string. Full Datavyu file name. Defaults to 'db'}

\item{out_dir}{A character string. Output directory. Defaults to \code{in_dir}.}

\item{out_fn}{A character string. Full output file name.}

\item{auto_write_over}{A logical value. If TRUE, new output file overwrites old.}

\item{code_regex}{A character string. Specifies the regular expression
to extract a code.}

\item{code_vals_regex}{A character string. Specifies the regular expression
to extract a code value.}

\item{code_type_regex}{A character string. Specifies the regular expression
for the code type.}

\item{vb}{A logical value. If TRUE, provides verbose output.}
}
\description{
Extract Codes From a Datavyu file.
}
\examples{
\donttest{
extract_codes()
}

}
