% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{import_column}
\alias{import_column}
\title{Import Datavyu column into R}
\usage{
import_column(column, folder = getOption("datavyur.folder"), asList = FALSE,
  append.colnames = TRUE, classes = getOption("datavyur.classlist"), ...)
}
\arguments{
\item{column}{The name of the column to import as used in the Datavyu .opf file}

\item{folder}{Character string corresponding to the folder path to be scanned. Defaults to option \code{datavyur.folder}.}

\item{asList}{Logical value indicating to return a list or data frame (default).}

\item{append.colnames}{If \code{true}, add column name to each argument, 
e.g., \code{column.arg}, instead of having column as a variable in the data.}

\item{classes}{List of new classes to override guessed classes when reading in .csv data. see \code{data.table::fread}}

\item{...}{Additional options passed to the function \code{data.table::fread}}
}
\description{
Imports a Datavyu column to R when using the datavyu2csv.rb script
}
\details{
This function only works if you had previously used the \code{datavyu2csv.rb} script to export a Datavyu file to .csv
This can be obtained from \url{http://github.com/iamamutt/datavyu/general}.
\cr
Note: If the same column name was used but has different number of arguments then you will get an error unless \code{asList=TRUE}.
This function assumes that the .csv is structured in a way based on how the \code{datavyu2csv.rb} script exports data.
}
\examples{
import_column("childhands")
import_column("childhands", append.colnames = TRUE)
}

