% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{merge_nested}
\alias{merge_nested}
\title{Merge nested data}
\usage{
merge_nested(c1, c2, folder = getOption("datavyur.folder"),
  classes = getOption("datavyur.classlist"), ...)
}
\arguments{
\item{c1}{top level column name as a character, (e.g., trial or block)}

\item{c2}{bottom level column name as a character (e.g., eye gazes within trial)}

\item{folder}{Character string of the name of the folder to be scanned. Defaults to option \code{datavyur.folder}.}

\item{...}{Additional options passed to \code{data.table::fread}}
}
\description{
Merges two Datavyu columns by onset/offset timestamps
}
\details{
Since data is nested, this will repeat rows from the higher level data.
If any cell in the lower level is not within a cell in the higher level, it
will not be in the final data set. This function is to be used only with truly
nested data. For finer control over potentially overlapping cells see 
\code{\link{temporal_align}} with the option \code{keep.frames=FALSE} and a high
value for \code{fps} for better precision in aligning timestamps.
}
\examples{
# merge nested data, throwing away lower level cells that are not within higher level cells
merged <- merged_nested("childhands", "parenthands")

# another way to merge nested, but contains NAs for cells that are not fully within one another
# since these data are not nested, you can see how much is being thrown away compared to x
y <- temporal_align(all.cols=c("childhands", "parenthands"), fps=60, keep.frames=FALSE)
}

