% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{ms2time}
\alias{ms2time}
\title{Convert milliseconds to a time string}
\usage{
ms2time(timestamp, unit = "\%H:\%M:\%S", msSep = ":")
}
\arguments{
\item{timestamp}{A numeric time duration, such as \code{1102013} ms}

\item{unit}{How the string will be contructed. Default is Hours:Minutes:Seconds.MS}

\item{msSep}{How the separator between seconds and ms will look like}
}
\description{
This will take a duration of time in milliseconds and convert it to a time string format
}
\examples{
# 18 minutes and 22 seconds and 13 milliseconds
ms2time(1102013)
}

