% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{ordinal_align}
\alias{ordinal_align}
\title{Ordinal alignment}
\usage{
ordinal_align(all.opf = TRUE, all.cols = TRUE,
  folder = getOption("datavyur.folder"),
  classes = getOption("datavyur.classlist"), ...)
}
\arguments{
\item{all.opf}{Use all found .opf files from \code{folder} or specify .opf file names to use. Must 
be a character vector with exact file name matches as seen inside one of the exported .csv files, which are 
just the original file names with .opf removed. If you choose to provide full path and file names, 
you MUST set \code{folder=NULL}.}

\item{all.cols}{Use all columns found in \code{folder} or only use specified column names entered
as a character vector.}

\item{folder}{Defaults to option \code{datavyur.folder}.}

\item{classes}{List of new classes to override guessed classes when reading in .csv data. see \code{data.table::fread}}

\item{...}{Additional arguments passed to \code{data.table::fread}, 
except \code{stringsAsFactors, colClasses, verbose, showProgress}}
}
\value{
A data.frame with merged data aligned by cell number (ordinal value)
}
\description{
Align and merge data by cell number (ordinal)
}
\details{
See \code{\link{temporal_align}} for more details on function usage.
}
\examples{
ordinal_align()
}
\seealso{
\code{\link{temporal_align}}
}

