% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfun.R
\name{ordinal_align}
\alias{ordinal_align}
\title{Ordinal alignment}
\usage{
ordinal_align(all.opf = TRUE, all.cols = TRUE,
  folder = getOption("datavyur.folder"),
  classes = getOption("datavyur.classlist"), ...)
}
\arguments{
\item{all.opf}{All .opf files (\code{TRUE}) or a character vector of specific .opf files to use}

\item{all.cols}{All columns from a .opf file or character vector of specific column names}

\item{folder}{Search folder, defaults to \code{datavyur.folder} option}

\item{classes}{Override column classes.}

\item{...}{Additional arguments passed to \code{data.table::fread}, 
except \code{stringsAsFactors, classes, verbose, showProgress}}
}
\value{
A data.frame with merged data aligned by cell number (ordinal value)
}
\description{
Align and merge data by cell number (ordinal)
}
\details{
See \code{\link{temporal_align}} for more details on function usage.
}
\examples{
ordinal_align()
}
\seealso{
\code{\link{temporal_align}}
}

