% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv2datavyu.R
\name{r2datavyu}
\alias{r2datavyu}
\title{R data to Datavyu .csv file}
\usage{
r2datavyu(rlist, filename = "datavyur_export")
}
\arguments{
\item{rlist}{List of columns to be used in the final Datavyu file.}

\item{filename}{Filename of the .csv file to be used}
}
\description{
Exports R data as a list or dataframe to a .csv file used by Datavyu for importing.
Each list item is a different column in the final datavyu file.
To import go to Datavyu and go to File > Import > import .csv
}
\examples{
# First get example data to use
example_data <- datavyu_dat()

# See how the example data is structured
str(example_data)

# Export R list to a .csv file for importing into Datavyu
r2datavyu(example_data, "example_file")
}

